tests/wpt/web-platform-tests/domparsing/outerhtml-01.html

<!DOCTYPE html>
<title>outerHTML: child of #document</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://html5.org/specs/dom-parsing.html#outerhtml">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws("NO_MODIFICATION_ALLOWED_ERR", function() {
    document.documentElement.outerHTML = "<html><p>FAIL: Should have thrown an error<\/p><\/html>";
  })
});
</script>
